requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
decorator==5.1.1
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-crontab==0.7.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-graphql-jwt==0.4.0
django-ipware==6.0.4
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
graphviz==0.20.1
h2==4.1.0
hpack==4.0.0
http-ece==1.2.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
ipython==8.17.2
jedi==0.19.1
lxml==5.1.0
macaddress==2.0.2
Markdown==3.5.1
matplotlib-inline==0.1.6
msgpack==1.0.7
netifaces==0.11.0
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
priority==1.3.0
promise==2.3
prompt-toolkit==3.0.41
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-vapid==1.9.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydotplus==2.0.2
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-generate-mac==1.3.1
python-ipware==2.0.1
pytz==2023.3.post1
pywebpush==1.9.4
PyYAML==6.0.1
randmac==0.1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.13.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.10
xlrd==2.0.1
zope.interface==6.1
